Extracts Matches modal into new component.
The rendering of the modal, when inline with the matches list was
causing the re-rendering of the whole list. When that list was large, the
rendering time was significant enough to slow down the page
interactions. This changes restricts the rendering to the modal, when it
is open, and greatly improves the page responsiveness.