Fixes match modal displaying wrong image.
The Select button in the match summary was displaying the modal
correctly, but with the wrong image. The reason was that the state was
defaulting to the first value the first time the modal was loaded, and
then to the value in the state on subsequent loads. The answer is to
initialize the state with the correct image index when the modal is
opened.