Moves Google Auth API init to index.html.
The Amazon code is expecting to be able to attach to the gapi object in
the global scope and fails when it can't. Moving the initialization to
the index.html might ensure that it is always available in the global
scope when the AWS code needs to try and refresh the tokens.