Adds media queries to improve homepage image
The background image on the home page was not positioned well, when the
screen size was reduced to something used on a cell phone. This fixes
the issue, by adding media queries to change the image to something more
subdued, so that the text can be read when on top of it.