Adds limit to # matched names used in search.
There are opportunities to generate a search query with wild cards that
will return 1000s of matched names from the list of all matches. We
don't want to load that many as it will tax both the browser when
loading the DOM and the AWS end points sending the data. Adding a limit
of 100 names should be fine.