Moves logged-in status to AppContext.
...and moves login button out of the navbar. The moving of the logged0in
status to the AppContext was prompted by the move of the logout button,
because that button was moved out of the App component. At which point,
it no longer had access to the local state functions to update the
logged-in status. Moving that status to the AppContext meant that we
could move the button and still access the functions required to log the
user out.