Converts Google to OAuth Redirect login.
The original login would authorize the google user, but not provide a
complete cognito user that is required to assign roles for access to the
s3 storage. Converting to the "HostedUI" allows us to generate a cognito
user. Adding the {provider: "Google"} option to the federatedSignIn()
call bypasses the hostedUI page and keeps the login flow looking the
same as before, with a few redirects.

Also had to upgrade Amplify to the latest version, which changed the way
the Storage calls return the data. Now instead of grabbing the body from
the object returned in the Promise from the get request we have to grab
the text, which returns another promise and then work with the object
returned to that promise.