Fixes mask selection.
There was a problem loading the image from AWs into the canvas to
generate a mask selection. This was only visible when then cache was
enabled, most users, so didn't get noticed until the code was out of
development. Ended up having to use the storage api to fetch the image
blob and build the image that way as there were too many CORS / taint
issues with the other methods.