Converts relative urls in results file to full url
The results files all have a relative url in them to make it easier to
move between versions of the data. These need to be converted back to
full urls by the website, when it tries to load the results from s3.
This update adds a fetch to get the paths.json file and uses that to
complete the url path before fetching the results or images.