Adds recursive fetch to graphQL query for searches
The default for graphQL queries on AppSync is to limit the number of
result returned to a low number, ~6. This means that without the
recursive query we only get the first 6 results back when querying the
backend. The recursive query continues to query the database until we
have all the results we are expecting. And now all searches are loaded
into the page.