Converts Search Upload to allow mask selection.
Prior to this the code assumed that if the image was aligned, then it
was also masked. It would then skip to the search step without asking
for any other input. Now the file is uploaded, a default channel is
created and the user is given the chance to create a mask for searching.