Adds alternative method to fetch items from S3
Admin users need an alternative method to fetch Items from S3, because
the code from Amplify (Storage.get) uses the currently logged in users
identityId to build the request to AWS when using the private level.
Since we want to be able to see other users files we need to pass their
identityId. In order to do this we have to use a nasty hack where we set
the query level to public and create a custom "private" prefix that
contains the identityId of the owner of the file we wish to get.

const storageOptions = {
  level: "public",
  download: true,
  customPrefix: { public: `private/${searchMeta.identityId}/` },
  bucket: config.SEARCH_BUCKET
};

Storage.get(url, storageOptions);