Re-imagines the mask selection process
The "Create Mask" button has been removed from the page. Selecting
nothing but a channel, will use the whole image as the mask and avoids
an upload. If the mouse is clicked within the mask drawing region, the
mask will be automatically created and the "Clear Mask" button will
become active. The user can then use the masked region and submit the
search or clear the mask and start again.