Sets ZIP_DOWNLOAD_LIMIT to 200.
This should prevent the lambda function from timing out before it can
zip all the files for download. If we want to increase the number of
images allowed or remove limits, then the lambda function will need to
be re-written to use dynamo db and sockets so that it can run async and
avoid the api gateway timeout.