Removes hard coded values from config.
The config had hard coded values for the cognito user pool and identity
pools. Using the serverless build and export env plugins, we can
generate those values during the build process and no longer need to
hard code them. This is also applied to the environment variables
provided in the .env file.