Changes .push to .replace for better UX.
Using the history.push command when modifying the url for a modal pop up
resulted in the modals reappearing when the back button was pressed.
This didn't seem like the correct behavior one would expect, so push has
been replaced with .replace to get the expected behavior when clicking
on the back button.