Moves filtering from FilterContext to querystring
Moving the filter parameters into the query string allows people to
share the page that they were looking at, including all the filters and
the modal pop up. It does not currently transfer the image selection on
the pop up as this is permanently set on the users local storage. There
is a chance that may have to change as well. This also makes it
difficult to preserve the filters between pages. There would need to be
a lot of work on all the links to keep the query state intact.