Crosshair mouse coordinates moved to context.
This improved performance and removed a lot of prop drilling. The
context is placed in the component tree above all the linked image and
it handles the coordinate updates. The components between the context
and the crosshair display don't have to update, because the mouse
position values are no longer passed to them.