Improves keyboard usability for search input.
The UI for the autocomplete of the search input was not working as
expected. Now when a value is selected in the autocomplete the search is
automatically triggered, rather than having to hit enter a second time,
or the worst case where hitting enter would search for the previously
typed value in the search input and not the selected value in the
autocomplete dropdown.