Wraps readMetaData return in a Promise
This makes it possible to wrap the entire data loading tree in a
Promise.all() function, so that we can be sure all the data has loaded
before running the cleanup code. Without this, the Top level promises
would resolve and the next layer of async code would execute after the
cleanup code.