Only include credentials when fetch url is signed.
If the include credentials header is used on an unsigned url, then the
S3 bucket will refuse to serve the file, even if it is public, if the
user is not in the allowed users list. This breaks the site when we
request images for buckets that are public and have no access
restrictions applied.