Fix pppm matches for Neuron Type/Instance searches
The previous version of the published names database was showing all
body ids as not having a match to pppm results, when located via a
search for a neuron type or instance. The updated version of the
database adds a true or false flag to each body id in the results to
indicate if pppm matches exist at the bodyid level. This needed to be
accounted for in the code.