Fixes displayed pppRank
We need to modify the ppp results to update the pppRank. The data
stored in the json files is the internal pppRank, which can have
half values (eg: 12.5). This does not match the rank values as
shown in the pdf files for external use. We are assured that the
ppp results are sorted by pppRank, so we can apply the index of the
array to the result to replace the pppRank for external display.