Adds isAdmin & disables maintenance pages for admins.
The isAdmin flag has been added to the AppContext for use on the site.
The first use is to turn of the maintenance pages for admins, so that we
can see the site in full during maintenance. This allows admins to check
the site is working during the maintenance process. It does not diable
the warning banner, so that we know the site is still inaccessible for
other users.