Adds image dimensions check for VNC vs brain
When someone uploads an image in a format supported by the browser, we
can determine the dimensions by loading it into the page. If it is
taller than it is wide, then we can guess that it is a vnc image and
modify the anatomical region input to select "VNC" instead of "Brain".
This guessing can't be performed on tiff, zip or any of the other file
formats, which can't be opened by the browser.