Restructured code to remove react act warnings
The ImageWithModal component was throwing the react act warning, because
it was nested inside other components and couldn't be wrapped directly
with the act function during testing. Moving it out of the components
and passing it in a child, means the parent component can be tested
without the warnings being triggered.