Prevents modal jump by setting image heights.
All the images in the image comparison modal would render as 0 height
until they had been loaded, causing the scroll position of the window to
jump back to the top. This was particularly annoying when looking at VNC
imagery and switching between matches. To prevent the jump back to the
top, we have to set the height of the images, so that the page render
can anticipate the loading of the image and save space for it, keeping
the page at a constant height.