Prevents blinking of image in ImagePlaceholder
When the same image is loaded into the page more than once, it was
blinking to the placeholder image. We only want the placeholder image to
show up if a new image is loaded. To fix that we check the path of the
image being loaded and if it matches the previous path, we leave the
loaded status set to true.