Removes anatomicalRegions from local config
Since this functionality is determined by the data, the configuration
was moved to S3 buckets with the data. The config.json associated with
each site now has an anatomicalRegions entry that can be used to
determine which sites are able to search which regions. This should
streamline the process of releasing new regions.