Moves disableAlignment setting to appContext.
This was originally set as an environment variable when releasing the
site. Since it is tied to the backend data and support, it is better to
have the setting configured in the config.json file retrieved from the
S3 bucket. Doing it that way makes it easier to enable the alignment
system in the UI, when the backend is updated.