Adds fileType information to FileUploadStep
The step on the searches summary page now shows if the uploaded file was
a 2D CDM or a 3D stack. The code does its best to guess, based on the
filetype extension, if the uploaded file was correct. If not it displays
an error icon with a tooltip explaining why the code thinks the image
was incorrect.