Fixes download name for custom search uploads.
When downloading the image that was uploaded to do a custom search, on
the search results page, the name of the downloaded file was not a valid
image name, due to the additional query parameters attached to the image
url. This change strips those off when creating the download name.