Fixes download issues in safari browser.
"cache-control": "no-cache" was added to the fetch options,
because safari was refusing to download images in a fetch request.
I created a proxy service that showed the requests weren't even being
sent to the 'signed' url unless the cache was disabled. This change
will force the image to be downloaded from the server, even if it was
once loaded into the http cache, but that seems like a small price to
pay for up to date and working downloads.