Pulls mirror button out of context menu for images
The mirror button was buried inside the context menu for each of the
images in the image comparison modal. This made it difficult to find and
meant that there was no immediate indication that the image being
observed was mirrored by default. Moving it out to a stand alone button
makes it more obvious that an image has been modified and gives an easy
way to restore it if required.