Disables gender filter if gender prop not found.
Before this change all results were filtered out when the gender
property was missing on the result set. If there are gender
attributes on the results object, then we can filter by gender,
if not then we need to skip the gender filter or no results will
be returned by default.