Changes search urls to point directly to matches
In the Meta components there are links to the precomputed matches for
the sample being displayed. Previously these links would go to the
search?q=<bodyid|line>. Now they go directly to the numbered match,
skipping a step where the user would have to find the search in a list
and click on the correct one.