Updates check for result buttons in SkeletonResult
The search result buttons now use the results information located in the
files section of the metadata for the image. This tells us if a search
was run, so we can add the buttons to the site. For example,
metaInfo?.files?.CDSResults shows us that the color depth search was
performed, so we can add that button.