Removes imageURL & thumbnailURL match attributes.
There were being created as a result object was loaded into the site,
but it obscured some issues when trying to get the correct file paths
for some of the images. All cases have been removed and we now use the
file object from each match to generate paths to the images.