Updates MatchReferences lookup to use new publishedNames
We are no longer relying on a translation table to convert the
libraryName to a dataset name when looking up publications that
reference an EM bodyid. Instead, the publishedName now includes the data
set and version concatenated to the body id, eg:
hembrain:v1.2.1:1234567. This id can be passed as is to the doi lookup
function, removing any potential errors in translation.