Scrolls page to top when matches grid is loaded
When the search page was really long and a result at the bottom was
clicked, it would load you in to the bottom of the matches grid. This is
almost never what is wanted, so a scroll to the top of the page was
added whenever the matches grid is loaded for the first time.