iMoves url path modification into data loaders.
There were a lot of places where the relative paths in the results and
image objects were concatenated with the prefixs from the configuration
file. In order to reduce the places where this code was found, it has
been consolidated into the utils lib. This makes it much easier to
change it when needed. Also changes the name of some of the prefixes
from ColorDpethMip... to CDM....