Sets maximum width on thumbnails in search steps.
There were cases where a really wide image was uploaded for use in a VNC
search. The website was showing the image as a really wide image and
overlapping some of the other content on the search steps display. This
sets a maximum width for "vertical" thumbnails, to keep them restricted
to the correct display area.