Changes positioning of "mirror" button.
The mirror button was causing the vertical VNC images to be out of
alignment when the image was mirrored, because the button was too wide
to fit on the same row as the image selection drop down. To keep all the
images aligned, the mirror buttons have been permanently moved to a
second row when the images are narrow and vertical.