Removes hoverable prop from summary cards.
This was showing a raised border around the card, which was good, but it
also changed the pointer to a hand, indicating it was a link, which was
bad. There is no obvious way to prevent the cursor change, so removing
the hoverable prop is the available option for now.