Adds delayed site reload to uploads list.
This reload happens every 10 minutes and is intended to make sure that
the site doesn't get stuck in a loading state, if the websockets are
closed from the other end. This can happen if the browser is left
waiting for too long or if the tab is not in the foreground.