Fixes missing image comparison after customization
If someone chose an image that was only available in one data set as
the first image to view and then looked at a different data set that
didn't have that image as an option, the image would be removed from the
page, with no option to replace it with something else. This update
keeps the image placeholder in the page and shows a message to let the
user know that the image selection is not available for the current data
set.