Fixes error message for unsupported image uploads.
Previously, when a user uploaded an image that was "already aligned",
but was not one of our supported file types, the site would show a
cryptic message and then delete the search record. This update preserves
the search and adds an error message to the record to help the user
figure out what to do next.