Improves body id/line name filter
The search now uses a regex instead of a word match, so it is more
relaxed on filtering input, or example it no longer needs to have the
complete body id such as 'hemibrain:v1.2.1:424025668' to filter out
the results, but can also have just '424025668'. The text associated
with the input has also been clarified to indicate that the search is
only looking at line names and body ids.