Patches search code to filter any results
Before this it would only filter if there were more than 1 results,
which means no filtering on the original search term would occur if
there was only one result. This allowed people to search for an em
dataset that doesn't exist and still get a match, eg:
'hemibrain:v4:362473525' would return 'hemibrain:v1.2.1:362473525' as a
match.