Adds wildcards to dataset search.
We want to be able to support links from neuprint, that are for datasets
not in neuronbridge. In order to do that without having to resort to
hard coding a dataset version translation table into one or the other
site, we can use dataset wildcards, eg: 'hemibrain:*' to search for all
versions of hemibrain.