fix: Resets selected page when filter applied.
When a filter is applied to the matches page, the selected page is reset
to 1, to prevent the site from showing no data when there are less
results than would be needed to get to the selected page. This is
probably the correct choice as keeping the same page selected would show
different data anyway.

Also added a check to see if page will be beyond the number of results
and redirect to page 1, for those people that like to monkey with url
parameters.