fix: prevents app crash on unified search error
The searchRegex code was running too soon and threw an error that was
being caught by the error boundary, which resulted in an infinite loop
of errors. Moving the code later in the execution allowed additional
checks to be run before this code executed. Those checks now force the
code to bail out before reaching the searchRegex code and prevent the
loop.