feat: adds ListItemsByOwner query and limits uploads list to 100 items
Adds a new graphql query that can be used to fetch the list of upload
searches. This speeds up the load times for this page. Adding a limit to
the fetchItemsNextToken function allows us to set a limit of 100
searches on the uploads page. Next step is to add a way to load searches
after the limit and then reduce that limit to ~10/20.