feat: Converts searchType to image collections
The custom search uploads will be run against image collections in the
future, so the upload form has been modified to provide a list of the
collections to choose from as opposed to choosing either EM or LM
imagery. This also adds a check to convert any older searches that have
been re run, so that they are using the correct collection.