fix: Allows searches without dataset version.
Previously only searches with an id or with a dataset:version:id would
return a result, due to some filtering that was occurring in the client.
It was requested that we also allow 'dataset:id' searches as well. This
fix removes the version check, when one is not provided. The search
results will now show all versions of a specific body id within a
dataset if the dataset version is not provided.