fix: Nasty hack to get image aspect correct in safari
The images in the summary were not rendering with the correct aspect
ratio when they were wide. This made them look like squashed squares,
instead of rectangles. I added another parameter to the ImagePlaceholder
component to fix this in the places where it was broken on safari,
because there was no general fix that worked for all the places that the
placeholder is called.