fix: removes incorrect "Additional matches ...." message
When performing a wildcard search, there are times when the Additional
matches in different datasets" message would appear when it was not
supposed to. This is because the search could return multiple hits for
the wildcard, eg LHPD2c7* would return a hit for LHPD2c7 and LHPD2c7_R,
these contain 4 and 2 bodyIds respectively, with two of the bodyIds
overlapping. The code filtered out the duplicates at a later stage, but
did not do so, before the check that displays the "Additional matches
...." message. This adds a check to remove all bodyIds that have been
seen before and prevents the message from being shown at the wrong time.