fix: Adds zoom icon to images that show larger image modal.
There are some images that show a modal of the image zoomed in and some
that show a detailed dataset, both had the link cursor and it was not as
clear which would show which. This adds a zoom icon to the images that
show a zoomed in image, to differentiate them from the other modals.